requirements.txt
Flask
Flask-Login
Flask-MySQLdb
Flask-WTF
Flask-Bcrypt
